project/backend/requirements.txt
asyncpg==0.23.0
fastapi==0.65.3
gunicorn==20.1.0
requests==2.25.1
tortoise-orm==0.17.4
uvicorn==0.14.0
cryptography==39.0.1
paramiko==3.0.0
python-multipart==0.0.6
SQLAlchemy==2.0.4
python-jose==3.3.0
passlib==1.7.4